ci: CI-adjustments (#381)
* ci(workflows): pin action versions

This commit uses the standard convention of specifying the versions
of GitHub actions in a semver comment after the SHA.
Tools like renovate will use this, and pick it up.

It will lead to better PR's from renovate, where instead of
getting PR's like
Currently diff: 2342342342 -> 344939345328
With this: v4.2.3 -> v4.2.4

Dont worry though, the SHA-hash is still connected to
the version. If not, Renovate will adjust/correct this.

Signed-off-by: Josef Andersson <josef.andersson@digg.se>

* build(codequality): add missing license check

Signed-off-by: Josef Andersson <josef.andersson@digg.se>

* ci(commit): extend description length

The default setting for the lint committer
(conform) is 72 for description length.

This will at times leads to fails, as for
example tools like Dependabot and Renovate
sometimes adds longer commits messages.
Bumping it to 92 in length, has worked fine
in other projects, so should be fine here to
even for those longer cases.

Signed-off-by: Josef Andersson <josef.andersson@digg.se>

---------

Signed-off-by: Josef Andersson <josef.andersson@digg.se>